navcontent updated
